EventsDemo: Events log as a separate file
